componentize read only objective
